Generate label data

+++ bert_pytorch/dataset/process_data/sample_walks.py
@@ -6,7 +6,7 @@ import argparse
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('-e', '--edgelist', type=str)
-    parser.add_argument('--number-walks', type=float, default=10)
+    parser.add_argument('--number-walks', type=int, default=10)
     parser.add_argument('--walk-length', type=float, default=20)
 
     args = parser.parse_args()
